refactor: Changed float -> Decimal for precise arithmetics

+++ calculator/calculator.py
@@ -1,39 +1,40 @@
 from calculator.calculation import Calculation
 from calculator.operations import add, subtract, multiply, divide
 from calculator.calculation_history import CalculationHistory
+from decimal import Decimal
 
 class Calculator:
     _history = CalculationHistory
 
     @staticmethod
-    def perform_operation(operation, a: float, b: float) -> float:
+    def perform_operation(operation, a: Decimal, b: Decimal) -> Decimal:
         calculation = Calculation(operation, a, b)
         Calculator._history.add_calculation(calculation)
         return calculation.perform_operation()
 
     @staticmethod
-    def add(a: float, b: float) -> float:
+    def add(a: Decimal, b: Decimal) -> Decimal:
         return Calculator.perform_operation(add, a, b)
 
     @staticmethod
-    def subtract(a: float, b: float) -> float:
+    def subtract(a: Decimal, b: Decimal) -> Decimal:
         return Calculator.perform_operation(subtract, a, b)
 
     @staticmethod
-    def multiply(a: float, b: float) -> float:
+    def multiply(a: Decimal, b: Decimal) -> Decimal:
         return Calculator.perform_operation(multiply, a, b)
 
     @staticmethod
-    def divide(a: float, b: float) -> float:
+    def divide(a: Decimal, b: Decimal) -> Decimal:
         """
         Divides a by b.
 
         Args:
-            a (float): Numerator.
-            b (float): Denominator.
+            a (Decimal): Numerator.
+            b (Decimal): Denominator.
 
         Returns:
-            float: The result of the division.
+            Decimal: The result of the division.
 
         Raises:
             ZeroDivisionError: If b is zero.
